test/regression/reftest/vendor/webkit/fast/dom/HTMLDocument/constructor-setter-crash.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/wpt-test-pre.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Make sure we don't crash when setting HTMLDocument.constructor");

shouldNotThrow("document.constructor = 'test'");

</script>
</body>
<script src="../../../resources/js-test-post.js"></script>
<script src="../../../resources/wpt-test-post.js"></script>
</html>
